# docs/swagger.yaml
basePath: /
definitions:
...
  categories.requestCreateStyle:
    properties:
      back:
        type: string
      background:
        type: string
      category_id:
        type: string
      front:
        type: string
      logo:
        type: string
      type:
        type: integer
    type: object
  categories.resAny:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
...
  credentials.Attribute:
    properties:
      id:
        type: integer
      name:
        type: string
      value:
        type: string
    type: object
  credentials.Data:
    properties:
      category:
        type: string
      description:
        type: string
      files:
        items:
          $ref: '#/definitions/credentials.File'
        type: array
      identifiers:
        items:
          $ref: '#/definitions/credentials.Identifier'
        type: array
      identity_number:
        type: string
      name:
        type: string
      type:
        type: integer
    type: object
  credentials.File:
    properties:
      file_encode:
        type: string
      id_file:
        type: integer
      name:
        type: string
    type: object
  credentials.Identifier:
    properties:
      attributes:
        items:
          $ref: '#/definitions/credentials.Attribute'
        type: array
      name:
        type: string
    type: object
  credentials.JwtTransactionRequest:
    properties:
      attributes_id:
        items:
          type: integer
        type: array
      block:
        type: integer
      files_id:
        items:
          type: integer
        type: array
      from:
        type: string
      to:
        type: string
      transaction_id:
        type: string
      ttl:
        type: integer
      verify:
        type: string
    type: object
  credentials.JwtTransactionResponse:
    properties:
      code:
        type: integer
      data:
        type: string
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
  credentials.ResGetFiles:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/credentials.File'
        type: array
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
  credentials.credential:
    properties:
      amount:
        type: number
      data:
        type: string
      from:
        type: string
      id:
        type: string
      to:
        type: string
      type_id:
        type: integer
    type: object
  credentials.requestCreateTransaction:
    properties:
      amount:
        type: number
      data:
        $ref: '#/definitions/credentials.Data'
      from:
        type: string
      to:
        type: string
      type_id:
        type: integer
    type: object
  credentials.responseAllCredentials:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/credentials.credential'
        type: array
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
  credentials.responseCreateCredential:
    properties:
      code:
        type: integer
      data:
        type: string
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
...
host: http://127.0.0.1:50050
info:
  contact:
    email: info@bjungle.net
    name: API Support
  description: Documentation Smart Contract
  license:
    name: Software Owner
    url: https://www.bjungle.net/terms/licenses
  termsOfService: https://www.bjungle.net/terms/
  title: OnlyOne Smart Contract
  version: "1.0"
paths:
...
  /api/v1/categories/create-style:
    post:
      consumes:
      - application/json
      description: create credential's style of category
      parameters:
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: request of create credential's style
        in: body
        name: CreateStyleOfCredential
        required: true
        schema:
          $ref: '#/definitions/categories.requestCreateStyle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/categories.resAny'
      summary: create credential's style of category
      tags:
      - Categories
  /api/v1/credentials/all/{block_id}/{limit}/{offset}:
    get:
      consumes:
      - application/json
      description: Get All Credentials ny block id
      parameters:
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Block ID
        in: path
        name: block_id
        required: true
        type: integer
      - description: Limit of pagination
        in: path
        name: limit
        required: true
        type: integer
      - description: Salt of pagination
        in: path
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credentials.responseAllCredentials'
      summary: Get all Credentials by block id
      tags:
      - Credentials
  /api/v1/credentials/create:
    post:
      consumes:
      - application/json
      description: Create credential
      parameters:
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request create transaction
        in: body
        name: createCredential
        required: true
        schema:
          $ref: '#/definitions/credentials.requestCreateTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credentials.responseCreateCredential'
      summary: Create credential
      tags:
      - Credentials
  /api/v1/credentials/files/{trx}:
    get:
      consumes:
      - application/json
      description: Get transaction files
      parameters:
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: transaction id
        in: path
        name: trx
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credentials.ResGetFiles'
      summary: Get transaction files
      tags:
      - Credentials
  /api/v1/credentials/jwt:
    post:
      consumes:
      - application/json
      description: Get JWTTransaction By ID
      parameters:
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Generate JWT Request
        in: body
        name: getJWTTransaction
        required: true
        schema:
          $ref: '#/definitions/credentials.JwtTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/credentials.JwtTransactionResponse'
      summary: Get JWTTransaction By ID
      tags:
      - Credentials
...
swagger: "2.0"
...